Make the crawler the main CLI
This shuffles things around to make the specs crawler the main part of Reffy
and analysis tools secondary tools on their way out to a separate companion
tool.

Main updates:
- A new `crawl-and-study.js` file was created out of the previous `reffy.js`
CLI. Relevant scripts in the `package.json` have been updated to target that
new tool accordingy, which should typically make the update transparent for
Webref.
- The "lib" bits of `crawl-specs.js` now appear in `src/lib/specs-crawler.js`
- The old `reffy.js` script now features the CLI bits of `crawl-specs.js`.
- The `reffy.js` is now the only binary exposed in `package.json`.
- `crawl-specs.js` was dropped accordingly.
- README adjusted to put the specs crawler first and clarifies that analysis
tools will go.

In other words, running `node reffy.js` now runs the crawler itself without
running the analysis tools. This is a step towards splitting the crawler
and the analysis tools (#747)

Depending on how external projects may have been using Reffy, typically
depending on whether they were using `reffy.js` and/or other CLI tools directly,
this update is breaking. Hence the major version bump.
