+++ tests/reffy-package.js
@@ -32,8 +32,8 @@ describe("The npm package of Reffy", function () {
   });
 
   it("can crawl specs", async () => {
-    const clidir = path.join(tmpdir, 'node_modules', 'reffy', 'src', 'cli');
-    const { crawlList } = require(path.join(clidir, 'crawl-specs'));
+    const clidir = path.join(tmpdir, 'node_modules', 'reffy', 'src', 'lib');
+    const { crawlList } = require(path.join(clidir, 'specs-crawler'));
     const refResults = JSON.parse(fs.readFileSync(__dirname + "/crawl-test.json", "utf-8"));
     const results = await crawlList(specs);
     for (const result of results) {
